All error and warnings resolved, and also attendance folder structure optimised :)
